Move all NPC components into a root NPCs folder.
