+++ src/components/npcs/npc-list/npc-list.ts
@@ -1,29 +1,23 @@
 import html from "./npc-list.html";
-import { IUnitOfWork } from "../../lib/common/data-access/unit-of-work-interface";
-import { appInjector } from "../../lib/infrastructure/app-injector";
-import { UnitOfWork } from "../../lib/data-access/unit-of-work";
-import { GetAllNpcsFeature } from "../../features/npcs/get-all-npcs/get-all-npcs-feature";
-import { NpcListItem } from "../../features/npcs/npc-list-item";
-import { EmptyRequest } from "../../lib/common/features/empty-request";
-import { EventBus } from "../../lib/events/event-bus";
-import { AppErrorEvent } from "../../lib/events/app-error-event";
-import { EventType } from "../../lib/events/event-type";
-import { AppEvent } from "../../lib/events/app-event";
-import { ModalDialogComponent } from "../modal-dialog/modal-dialog";
+import { GetAllNpcsFeature } from "../../../features/npcs/get-all-npcs/get-all-npcs-feature";
+import { NpcListItem } from "../../../features/npcs/npc-list-item";
+import { EmptyRequest } from "../../../lib/common/features/empty-request";
+import { EventBus } from "../../../lib/events/event-bus";
+import { EventType } from "../../../lib/events/event-type";
+import { AppEvent } from "../../../lib/events/app-event";
+import { ModalDialogComponent } from "../../modal-dialog/modal-dialog";
 import { NpcFormComponent } from "../npc-form/npc-form";
-import { Npc } from "../../features/npcs/npc";
-import { GetNpcByIdFeature } from "../../features/npcs/get-npc-by-id/get-npc-by-id-feature";
-import { GetNpcByIdRequest } from "../../features/npcs/get-npc-by-id/get-npc-by-id-request";
-import { SortNpcsListFeature } from "../../features/npcs/sort-npcs-list/sort-npcs-list-feature";
-import { SortState } from "../../lib/sorting/sort-state";
-import { SortNpcsListRequest } from "../../features/npcs/sort-npcs-list/sort-npcs-list-request";
-import { SortDirection } from "../../lib/sorting/sort-direction";
-import { TableHeader } from "../../lib/tables/table-header";
-import { NpcFilterChangedEvent } from "../../lib/events/npc-filter-changed-event";
-import { AppEventListener } from "../../lib/events/app-event-listener-interface";
-import { FilterNpcsListFeature } from "../../features/npcs/filter-npcs-list/filter-npcs-list-feature";
-import { FilterNpcsListRequest } from "../../features/npcs/filter-npcs-list/filter-npcs-list-request";
-import { BaseListComponent } from "../base-list/base-list-component";
+import { Npc } from "../../../features/npcs/npc";
+import { GetNpcByIdFeature } from "../../../features/npcs/get-npc-by-id/get-npc-by-id-feature";
+import { GetNpcByIdRequest } from "../../../features/npcs/get-npc-by-id/get-npc-by-id-request";
+import { SortNpcsListFeature } from "../../../features/npcs/sort-npcs-list/sort-npcs-list-feature";
+import { SortNpcsListRequest } from "../../../features/npcs/sort-npcs-list/sort-npcs-list-request";
+import { TableHeader } from "../../../lib/tables/table-header";
+import { NpcFilterChangedEvent } from "../../../lib/events/npc-filter-changed-event";
+import { AppEventListener } from "../../../lib/events/app-event-listener-interface";
+import { FilterNpcsListFeature } from "../../../features/npcs/filter-npcs-list/filter-npcs-list-feature";
+import { FilterNpcsListRequest } from "../../../features/npcs/filter-npcs-list/filter-npcs-list-request";
+import { BaseListComponent } from "../../base-list/base-list-component";
 
 export class NpcListComponent extends BaseListComponent {
     private npcsList: Array<NpcListItem>;
